fix(ios): extract BGContinuedProcessingTask code into its own file
WorkmanagerPlugin.swift exceeded SwiftLint's 1000-line file_length limit
(1125 lines) after adding the BGContinuedProcessingTask API surface, which
failed format_swift. Move the continued-processing registration, scheduling
and handling code into BGContinuedProcessingTaskScheduler.swift, mirroring
the existing one-file-per-concern layout (BackgroundWorker,
MacOSActivityScheduler, etc.). Widen four helpers from private to internal
so the cross-file extension can call them; the module's public API is
unchanged.
